Follow up with adjustments to Windows JobObject implementation
