+++ x-pack/osquerybeat/internal/osqd/osqueryd.go
@@ -21,9 +21,8 @@ import (
 	"github.com/dolmen-go/contextio"
 	"github.com/pkg/errors"
 
-	"github.com/elastic/beats/v7/libbeat/common/proc"
 	"github.com/elastic/beats/v7/libbeat/logp"
-
+	"github.com/elastic/beats/v7/x-pack/libbeat/common/proc"
 	"github.com/elastic/beats/v7/x-pack/osquerybeat/internal/fileutil"
 )
